code refactorying
