+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group = 'com.bytmasoft'
-version = '0.0.1-SNAPSHOT'
+version = '1.0.0'
 
 java {
     toolchain {
