+++ gradle.properties
@@ -1,9 +1,12 @@
 mapstructVersion=1.5.5.Final
 lombokVersion=1.18.30
 lombokMapstructBindingVersion=0.2.0
+commonVersion=1.0.9
 securityVersion=1.0.0
 dotenvVersion=4.0.0
 flywayVersion=9.0.0
 openApiVersion=2.3.0
-commonVersion=1.0.7
+jjwtVersion=0.11.5
+logstashVersion=7.4
+
