+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'document-management-service'
+rootProject.name = 'dm-service'
